feat: added errors to media routes

+++ src/medias/medias.service.ts
@@ -1,6 +1,5 @@
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateMediaDto } from './dto/create-media.dto';
-import { UpdateMediaDto } from './dto/update-media.dto';
 import { MediasRepository } from './medias.repository';
 import { Media } from '@prisma/client';
 
@@ -9,8 +8,14 @@ export class MediasService {
   private medias: Media;
   constructor(private readonly repository:MediasRepository) {
   }
+
   async create(body: CreateMediaDto) {
-    const { title, username } = body;
+    const medias = await this.repository.findAll();
+    for (const media of medias) {
+      if (media.title === body.title && media.username === body.username) {
+        throw new ConflictException("This user/title combination already exists!"); // Encontrou uma correspondência
+      }
+    }
     return await this.repository.create(body);
   }
 
@@ -19,14 +24,26 @@ export class MediasService {
   }
 
   async findMedia(id: number) {
-    return await this.repository.findMediaById(id);
+    const media = await this.repository.findMediaById(id);
+    if(!media) throw new NotFoundException("Media not found!")
+    return media;
   }
 
   async update(id: number, body: CreateMediaDto) {
+    const media = await this.repository.findMediaById(id);
+    if(!media) throw new NotFoundException("Media not found!")
+    const medias = await this.repository.findAll();
+    for (const media of medias) {
+      if (media.title === body.title && media.username === body.username) {
+        throw new ConflictException("This user/title combination already exists!"); // Encontrou uma correspondência
+      }
+    }
     return await this.repository.update(id, body);
   }
 
   async remove(id: number) {
+    const media = await this.repository.findMediaById(id);
+    if(!media) throw new NotFoundException("Media not found!")
     return await this.repository.remove(id);
   }
 }
